+++ core/execution/spot/market_test.go
@@ -164,7 +164,6 @@ func getMarketWithDP(base, quote string, pMonitorSettings *types.PriceMonitoring
 				TimeWindow:    3600, // seconds = 1h
 				ScalingFactor: num.DecimalFromFloat(10),
 			},
-			TriggeringRatio: num.DecimalZero(),
 		},
 		LiquiditySLAParams: &types.LiquiditySLAParams{
 			PriceRange:                  num.DecimalFromFloat(0.05),